+++ src/main/java/gregtech/integration/jei/GTJeiPlugin.java
@@ -24,16 +24,9 @@ import gregtech.common.blocks.MetaBlocks;
 import gregtech.common.gui.widget.craftingstation.CraftingSlotWidget;
 import gregtech.common.items.MetaItems;
 import gregtech.common.metatileentities.MetaTileEntities;
-import gregtech.integration.jei.basic.GTFluidVeinCategory;
-import gregtech.integration.jei.basic.GTFluidVeinInfo;
-import gregtech.integration.jei.basic.GTOreCategory;
-import gregtech.integration.jei.basic.GTOreInfo;
+import gregtech.integration.jei.basic.*;
 import gregtech.integration.jei.multiblock.MultiblockInfoCategory;
 import gregtech.integration.jei.recipe.*;
-import gregtech.integration.jei.recipe.primitive.MaterialTree;
-import gregtech.integration.jei.recipe.primitive.MaterialTreeCategory;
-import gregtech.integration.jei.recipe.primitive.OreByProduct;
-import gregtech.integration.jei.recipe.primitive.OreByProductCategory;
 import gregtech.integration.jei.utils.MachineSubtypeHandler;
 import gregtech.integration.jei.utils.MetaItemSubtypeHandler;
 import gregtech.integration.jei.utils.ModularUIGuiHandler;
