+++ src/main/java/gregtech/integration/jei/basic/BasicRecipeCategory.java
@@ -1,4 +1,4 @@
-package gregtech.integration.jei.recipe.primitive;
+package gregtech.integration.jei.basic;
 
 import gregtech.api.GTValues;
 import mezz.jei.api.IGuiHelper;
@@ -14,7 +14,6 @@ import javax.annotation.Nullable;
 import java.util.Collections;
 import java.util.List;
 
-//TODO, move this and all inheritors into integration.jei.basic
 public abstract class BasicRecipeCategory<T, W extends IRecipeWrapper> implements IRecipeCategory<W>, IRecipeWrapperFactory<T> {
 
     public final String uniqueName;
